extract factory for tools
